+++ R/pp_summary.R
@@ -18,7 +18,7 @@
 #'   "anxiety =~ ", paste0("x", seq(1, 14, 2), collapse = " + "), "\n",
 #'   "depression =~ ", paste0("x", seq(2, 14, 2), collapse = " + ")
 #' )
-#' bmasem(
+#' fit <- bmasem(
 #'   model_syntax,
 #'   sample_cov = Norton13$data, sample_nobs = Norton13$n, orthogonal = TRUE
 #' )