+++ tests/testthat/test-norton_r.R
@@ -25,6 +25,7 @@ test_that("Corr: Dependent samples works for meta-CFA on Norton13", {
   )))
   bmasem_test_hist(fit)
   bmasem_test_trace(fit)
+  bmasem_test_comp_rel(fit)
   expect_error(
     print_out <- capture_output(pp_summary(fit), width = 300),
     NA