+++ R/bmasem.R
@@ -13,8 +13,7 @@
 #' @param sample_cov (list of matrices) sample covariance or correlation
 #' matrices.
 #' The rownames and/or colnames must contain the observed variable names.
-#' There may be missing elements in correlation matrices but not covariance
-#' matrices (for now).
+#' For now, assumes there are no missing elements in the covariance matrices.
 #' @param sample_nobs (vector of positive integer) Number of observations
 #' for each study.
 #' @param correlation (LOGICAL)
@@ -47,9 +46,6 @@
 #' Only relevant for analysis of correlation structures.
 #' If TRUE, sample levels of the study-level random effect (usually faster);
 #' If FALSE, don't.
-#' If there are missing elements in any correlation matrix,
-#' the choice here is overriden and set to FALSE.
-#' Without this constraint, model estimation may not converge.
 #' @param seed (positive integer) seed, set to obtain replicable results.
 #' @param warmup (positive integer) The number of warmup iterations to run per
 #' chain.
@@ -169,12 +165,6 @@ bmasem <- function(
       do.fit = FALSE, orthogonal = orthogonal
     )
   } else {
-    if (isTRUE(correlation)) {
-      sample_cov <- lapply(sample_cov, \(x) {
-        suppressWarnings(x <- stats::cov2cor(x))
-        x
-      })
-    }
     lav_fit <- lavaan::cfa(
       model,
       sample.cov = sample_cov, sample.nobs = sample_nobs, std.lv = TRUE,