+++ R/bmasem.R
@@ -44,7 +44,7 @@
 #' This feature is experimental, see details below.
 #' @param conditional_re (LOGICAL)
 #' Only relevant for analysis of correlation structures.
-#' If TRUE, sample levels of the study-level random effect;
+#' If TRUE, sample levels of the study-level random effect (usually faster);
 #' If FALSE, don't.
 #' @param seed (positive integer) seed, set to obtain replicable results.
 #' @param warmup (positive integer) The number of warmup iterations to run per